investment-team: extract AST helpers + zero-trade repair + per-call RuleContext
Three structural reductions on top of the rule-registry pass:

1. **Extract code_safety AST helpers** → ``code_safety_ast.py``.
   ``code_safety.py`` drops from 1001 → 373 lines: ~600 lines of
   AST walkers (alias tracking, closure descent, universe-guard
   pattern matching) move into a sibling module so the gate file
   focuses on rules and the registry, and the intricate AST analysis
   gets its own testable boundary.

2. **Extract the zero-trade repair sub-pipeline** → ``zero_trade_repair.py``.
   ``orchestrator.py`` drops by ~390 lines: ``_run_zero_trade_repair``
   (~340 lines) plus ``_apply_zero_trade_spec_updates`` and the
   ``_ZeroTradeRepairOutcome`` dataclass move into a dedicated
   ``ZeroTradeRepairer`` class. The orchestrator now delegates with
   ``self.zero_trade_repairer.try_repair(...)`` at the single call site.

3. **Replace ``self._<call_state>`` slots with frozen ``RuleContext``
   dataclasses** for ``SpecReadinessGate`` / ``CodeSafetyChecker`` /
   ``BacktestAnomalyDetector``. Each rule's signature now declares its
   dependencies explicitly (``def _check_X(self, ctx: <Gate>Ctx)``);
   the ``try/finally`` cleanup of per-call state attributes is gone,
   and concurrent ``check()`` calls are safe.

Test surface: zero-trade tests updated to import from the new module
(``_ZeroTradeRepairOutcome`` aliased to ``ZeroTradeRepairOutcome``)
and monkeypatch ``run_strategy_code`` on both ``orchestrator_module``
and ``zero_trade_repair_module`` since the repair flow now imports it
directly.
